Cleaning cluster api from click imports
