+++ python/fedml/cli/modules/cluster.py
@@ -2,6 +2,8 @@ import click
 
 import fedml.api
 
+from prettytable import PrettyTable
+
 
 @click.group("cluster")
 @click.help_option("--help", "-h")
@@ -49,7 +51,14 @@ def kill(version, api_key, cluster_names):
     help="specify version of MLOps platform. It should be dev, test or release",
 )
 def killall(version, api_key):
-    fedml.api.killall_clusters(version=version, api_key=api_key)
+    cluster_list_obj = fedml.api.list_clusters(version=version, api_key=api_key)
+    if cluster_list_obj and cluster_list_obj.cluster_list:
+        _print_clusters(cluster_list_obj)
+        if click.confirm(f"Are you sure you want to kill all these clusters?", abort=False):
+            fedml.api.killall_clusters(version=version, api_key=api_key)
+    else:
+        click.echo("No clusters found.")
+
 
 
 @fedml_clusters.command("list", help="List clusters from the MLOps platform.")
@@ -66,4 +75,18 @@ def killall(version, api_key):
     help="specify version of MLOps platform. It should be dev, test or release",
 )
 def list_clusters(version, api_key, cluster_names):
-    fedml.api.list_clusters(version=version, api_key=api_key, cluster_names=cluster_names)
+    cluster_list_obj = fedml.api.list_clusters(version=version, api_key=api_key, cluster_names=cluster_names)
+    if cluster_list_obj and cluster_list_obj.cluster_list:
+        _print_clusters(cluster_list_obj)
+    else:
+        click.echo("No clusters found.")
+
+
+def _print_clusters(cluster_list_obj):
+    click.echo("Found the following matching clusters.")
+    cluster_list_table = PrettyTable(['Cluster Name', 'Cluster ID', 'Status'])
+
+    for cluster in cluster_list_obj.cluster_list:
+        cluster_list_table.add_row([cluster.cluster_name, cluster.cluster_id, cluster.status])
+
+    print(cluster_list_table)
